rules: Rename to `stackableShulkerBoxesInInventories` and register in MixinConfig
