+++ src/main/java/quickcarpet/settings/Settings.java
@@ -309,7 +309,7 @@ public class Settings {
     public static boolean stackableShulkerBoxes = false;
 
     @Rule(category = SURVIVAL)
-    public static boolean stackShulkerBoxesInInventories = false;
+    public static boolean stackableShulkerBoxesInInventories = false;
 
     @Rule(category = {FEATURE, CREATIVE})
     public static boolean terracottaRepeaters = false;
