+++ src/main/java/quickcarpet/utils/MixinConfig.java
@@ -68,7 +68,7 @@ public class MixinConfig {
         .putAll("sparkingLighter", "sparkingLighter")
         .putAll("spawnChunkLevel", "spawnChunkLevel")
         .putAll("spawning", "shulkerSpawningInEndCities", "phantomsRespectMobcap")
-        .putAll("stackableShulkerBoxes", "stackableShulkerBoxes")
+        .putAll("stackableShulkerBoxes", "stackableShulkerBoxes", "stackableShulkerBoxesInInventories")
         .putAll("terracottaRepeaters", "terracottaRepeaters")
         .putAll("tileTickLimit", "tileTickLimit")
         .putAll("tnt", "explosionNoBlockDamage", "tntUpdateOnPlace", "tntPrimeMomentum", "tntHardcodeAngle")
